1. DTOs represented. 2. Jsons now passed to jsp. 3. Parser changed - made working on Mac. 4. config -> resource directory renaming (more correct maven naming).

+++ src/main/java/pl/ktif/ttlz/dao/TournamentDAO.java
@@ -5,10 +5,10 @@ import org.springframework.stereotype.Repository;
 import pl.ktif.ttlz.model.Tournament;
 
 @Repository
-public interface TournamentDAO extends org.springframework.data.jpa.repository.JpaRepository<Tournament, Long> {
+public interface TournamentDAO extends org.springframework.data.jpa.repository.JpaRepository<Tournament, Integer> {
 
 	Tournament findByName(final String name);
 
-	Tournament findById(final Long id);
+	Tournament findById(final Integer id);
 
 }
